function/requirements.txt
adal==1.2.3
applicationinsights==0.11.10
argcomplete==1.12.3
astroid==2.4.1
azure-cli-core==2.34.1
azure-cli-telemetry==1.0.6
azure-common==1.1.25
azure-core==1.26.1
azure-cosmos==4.2.0
azure-functions==1.2.1
azure-mgmt-core==1.3.2
azure-servicebus==0.50.2
bcrypt==4.0.1
bson==0.5.10
certifi==2020.4.5.1
cffi==1.14.0
chardet==3.0.4
click==7.1.2
colorama==0.4.3
cryptography==38.0.3
distlib==0.3.6
dominate==2.5.1
feedgen==0.9.0
filelock==3.8.0
Flask==1.1.2
Flask-Bootstrap==3.3.7.1
Flask-SQLAlchemy==2.4.1
humanfriendly==10.0
idna==2.9
isodate==0.6.0
isort==4.3.21
itsdangerous==1.1.0
Jinja2==2.11.2
jmespath==1.0.1
knack==0.9.0
lazy-object-proxy==1.4.3
lxml==4.9.1
MarkupSafe==1.1.1
mccabe==0.6.1
msal==1.20.0
msal-extensions==0.3.1
msrest==0.6.13
msrestazure==0.6.4
oauthlib==3.1.0
packaging==21.3
paramiko==2.12.0
pipenv==2022.11.5
pkginfo==1.8.3
platformdirs==2.5.3
portalocker==1.7.1
psutil==5.9.3
psycopg2-binary==2.9.4
pycparser==2.20
Pygments==2.13.0
# PyJWT==1.7.1
pylint==2.5.2
pymongo==3.10.1
PyNaCl==1.5.0
pyOpenSSL==22.0.0
pyparsing==3.0.9
pyreadline3==3.4.1
PySocks==1.7.1
python-dateutil==2.8.1
python-http-client==3.2.7
PyYAML==6.0
requests==2.23.0
requests-oauthlib==1.3.0
sendgrid==6.3.0
six==1.14.0
SQLAlchemy==1.3.16
tabulate==0.9.0
toml==0.10.0
typing_extensions==4.4.0
uamqp==1.2.7
urllib3==1.25.9
virtualenv==20.16.6
virtualenv-clone==0.5.7
visitor==0.1.3
Werkzeug==1.0.1
wrapt==1.12.1
